Restructure page folder
