+++ src/pages/SignInPage/index.css
@@ -1,0 +1,8 @@
+.SignIn__container {
+  margin-top: 3em;
+  text-align: left;
+}
+
+.SignIn_button {
+  margin-top: 12px;
+}
