+++ src/pages/ToDoPage/index.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useReducer, useRef, useState} from 'react';
 import {RouteComponentProps} from 'react-router-dom';
 
-import reducer, {initialState} from './store/reducer';
+import reducer, {initialState} from '../../store/reducer';
 import {
     setTodos,
     createTodo,
@@ -9,10 +9,12 @@ import {
     toggleAllTodos,
     deleteAllTodos,
     updateTodoStatus
-} from './store/actions';
-import Service from './service';
-import {TodoStatus} from './models/todo';
-import {isTodoCompleted} from './utils';
+} from '../../store/actions';
+import Service from '../../service';
+import {TodoStatus} from '../../models/todo';
+import {isTodoCompleted} from '../../utils';
+
+import './index.css'
 
 type EnhanceTodoStatus = TodoStatus | 'ALL';
